requirements.txt
numpy
pillow
customtkinter
scipy
scikit-image
rasterio
matplotlib
numba
pyproj